Adding a stormy.py script.
Right now all this does is allow users to run `configure`, which will initialize
a ~/.stormy file with the user's API credentials.

Now we can start doing the fun stuff: building the actual CLI commands that
users will interact with.

NOTE: I'm using Docopt because it rocks, and is super simple.